Farm and FarmTest
FarmTest is a little sketch since I tested multiple things by editing the same test method multiple times. Farm displays the right values but not in an organized manner. I need to fix that after you lemme know how the GUI needs to values to be inputted to it. LMK if you have questions; the code isn't the neatest.