Allow more inputs for the month field in GUI
Previously, the only valid input for the month field was the 'MM'
format, where the month is two digits e.g. "05" for May.

However, this was not very intuitive and if the user entered the month
in the wrong format (e.g. "5" for May), it was not obvious that it was
the wrong format.

This makes it so the user can specify the months using a wider variety
of formats -- May can be specified as "05" (as before), simply "5", or
"May".